Use OpenGL Core Profile 4.1 for rendering
Kept the shaders as-is, only added the now mandatory vertex array. Using a more recent GL profile allows render pipeline debugging with RenderDoc and the use of more modern shading language features.

Should be available on all hardware/OSs in use today, as the standard was released in 2010. This is the highest OpenGL Standard supported by macOS.