Outsource renderer(s) as shared libraries
This will allow loading the renderers one by one and try them out, while not linking any 3D API libraries to the main executable. This will allow the game to run even if a required library is not available on the current system, e.g. libvulkan.so on Linux. In this case, loading the renderer lib will fail due to the dynamic linker failing to resolve the required symbols, and the renderer can be skipped.